feat: refactor election details structure, add race and candidate pages, and implement redirect handling
