+++ app/composables/useElectionByFfId.ts
@@ -1,0 +1,11 @@
+export const useElectionByFfId = async (id: string) => {
+  const { data: election, status: electionStatus } = await useAsyncData(
+    `election-${id}`,
+    () =>
+      queryCollection("elections")
+        .where("fiveFifthsId", "=", id)
+        .first(),
+  );
+
+  return { election, electionStatus };
+};
