+++ app/pages/journey/election/[electionId]/candidates/score/[id].vue
@@ -1,11 +1,14 @@
 <script setup lang="ts">
-const route = useRoute();
-const electionId = route.params.electionId as string;
-const candidateId = route.params.id as string;
+definePageMeta({
+  middleware: (to) => {
+    const electionId = to.params.electionId as string;
+    const candidateId = to.params.id as string;
 
-// Redirect old bookmark to new route with raceId 'r'
-await navigateTo(`/journey/election/${electionId}/candidates/r/score/${candidateId}`, {
-  redirectCode: 301,
+    return navigateTo(
+      `/journey/election/${electionId}/candidates/r/score/${candidateId}`,
+      { redirectCode: 301 },
+    );
+  },
 });
 </script>
